renamed folders, updated README and requirements.
