+++ python-sdk/utils/geometry_utils.py
