+++ python-sdk/utils/map_mask.py
