+++ requirements.txt
@@ -1,4 +1,3 @@
-easydict==1.8
 jupyter==1.0.0
 matplotlib==2.2.3
 numpy==1.14.5
@@ -6,4 +5,3 @@ opencv-python==3.4.2.17
 Pillow==5.2.0
 pyquaternion==0.9.2
 tqdm==4.25.0
-opencv-python==3.4.2.17
